remove some arguments that can also be fed at the caller side
